requirements.txt
disnake==2.9.1
python-dotenv==1.0.0
mafic==2.10.1
Pillow==10.2.0
aiohttp==3.9.3
requests==2.31.0
PyNaCl==1.5.0 